Script python module install for TA
Created a script that downloads and installs the latest python modules for
kintyre-speedtest-TA.  Module are installed into the target directory bin/lib,
and a requirements.txt file is written that can be used for external purposes
via 'pip install -r requirements.txt', otherwise this file is just use for
tracking the current versions of the packages that ship with the TA.

Updated the kintyre-speedtest.py scripted input to work with newer upstream
versions of the kintyre-speedtest-agent script (0.3.5 and later) which will be
upgraded in a following commit.  This also adds support for fetching the UUID
(or GUID) from the splunk instance configuration file to mirror what the
standalone agent does.